feat(marketplace): agents buy assets and trade coins from their own wallets
Extends the agent-wallet payment work from x402 checkout to two more rails:

- Asset purchases (avatars, agents, plugins): POST /api/marketplace/buy-asset
  accepts an agent_id and settles server-side from that agent's custodial
  wallet, verifying the transfer on-chain before granting via the same
  finalize path the browser flow uses. The purchase modal lists owned agents
  with live USDC balances; underfunded ones render disabled.
- Coin trading: the in-world coin widget gets a Pay with selector routing
  SOL-denominated buys and sells through /api/agents/:id/solana/trade, which
  already carried the shared spend guardrails but had no UI.

New _lib/agent-purchase.js unifies the daily autonomous-purchase cap across
skill_purchases and asset_purchases, so budget spent on one is no longer
available again for the other, and purchase-as-agent now shares it.

+++ api/_lib/agent-purchase.js
@@ -4,7 +4,7 @@
  *
  * Two marketplace surfaces sell things an agent can buy for itself: individual
  * skills (`skill_purchases`, via /api/marketplace/purchase-as-agent) and whole
- * assets — avatars, agents, plugins (`asset_purchases`, via
+ * assets (avatars, agents, plugins; `asset_purchases`, via
  * /api/marketplace/buy-asset with an `agent_id`). Both spend real USDC out of the
  * same wallet under the same owner-configured ceiling, so the cap accounting,
  * keypair recovery and transfer construction live here rather than being forked
@@ -16,14 +16,14 @@
  * assets. Enforcement is two-phase at every call site, and the second phase is
  * the authoritative one:
  *
- *   1. Pre-check before doing expensive work — a cheap early rejection, and a
+ *   1. Pre-check before doing expensive work: a cheap early rejection, and a
  *      TOCTOU on its own (concurrent buys all read the same pre-spend SUM).
  *   2. Re-check AFTER the pending row is inserted. The row is now counted by the
  *      same SUM, so every concurrent in-flight purchase is visible and the total
  *      cannot be collectively overshot. Fails safe: contending requests abort
  *      rather than risk exceeding the owner's limit.
  *
- * Nothing here decides WHETHER a purchase is allowed beyond the cap — ownership,
+ * Nothing here decides WHETHER a purchase is allowed beyond the cap. Ownership,
  * rate limits, self-dealing and already-owned checks stay with the endpoints,
  * which have the surface-specific context to answer them.
  */
@@ -42,7 +42,7 @@ import { ensureAgentWallet, recoverSolanaAgentKeypair } from './agent-wallet.js'
 
 export const USDC_DECIMALS = 6;
 
-/** Start of the current UTC day, as an ISO string — the cap window boundary. */
+/** Start of the current UTC day, as an ISO string: the cap window boundary. */
 export function dayStartIso() {
 	const d = new Date();
 	d.setUTCHours(0, 0, 0, 0);
@@ -51,7 +51,7 @@ export function dayStartIso() {
 
 /**
  * Read the owner-configured autonomous purchase ceiling off an agent's meta.
- * @param {object|null} meta — agent_identities.meta
+ * @param {object|null} meta - agent_identities.meta
  * @returns {{ enabled: boolean, limitUsdc: number|null, limitAtomics: bigint }}
  */
 export function readPurchaseCap(meta) {
@@ -92,8 +92,10 @@ export async function sumDailyPurchaseAtomics({ userId, currencyMint }) {
 				  AND currency_mint = ${currencyMint}
 				  AND status NOT IN ('failed', 'expired')
 			), 0)
-		)::numeric AS total
+		)::bigint AS total
 	`;
+	// ::bigint (not ::numeric) is load-bearing: SUM() yields numeric, which the
+	// driver can hand back as "2500000.00", a string BigInt() refuses to parse.
 	return BigInt(row?.total ?? 0);
 }
 
@@ -110,7 +112,7 @@ export function capExceededMessage(limitUsdc) {
  * missing or corrupt. Every decrypt is audit-logged with `reason` into
  * usage_events + the owner-viewable custody trail.
  *
- * The caller must ALREADY have verified that `userId` owns `agentId` — this
+ * The caller must ALREADY have verified that `userId` owns `agentId`; this
  * function does not authorize, it only loads.
  *
  * @param {{ agentId: string, userId: string, reason: string, meta?: object }} args
